chore: merge 후 파일 정리
