+++ apps/reflections/api_urls.py
@@ -1,11 +1,4 @@
-from django.urls import path, include
-from rest_framework.routers import DefaultRouter
-from .views import TeamViewSet, TeamMemberViewSet
-
-router = DefaultRouter()
-router.register(r'teams', TeamViewSet, basename='team')
-router.register(r'team-members', TeamMemberViewSet, basename='team-member')
+from django.urls import path
 
 urlpatterns = [
-    path('', include(router.urls)),
 ]
